requirements-gpu.txt
tensorflow-gpu
scikit-learn
matplotlib
Pillow
tqdm
gdown
nltk
seaborn
pandas
imageio

# conda install -c anaconda cudatoolkit=10.1
# conda install -c anaconda cudnn
# conda install notebook ipykernel nb_conda_kernels
# pip install git+https://github.com/keras-team/keras-tuner.git
# pip install autokeras